+++ src/model/Model.h
@@ -71,7 +71,7 @@ protected:
     Environment environment;
     Observations observations;
     int maxServers;
-    double chaosCoefficient;
+    int chaosCoefficient;
     double bootDelay;
     int horizon;
     int serverThreads;
@@ -115,7 +115,7 @@ public:
     const Observations& getObservations() const;
     void setObservations(const Observations& observations);
     int getMaxServers() const;
-    double getChaosCoefficient() const;
+    int getChaosCoefficient() const;
     double getEvaluationPeriod() const;
     double getBootDelay() const;
     int getHorizon() const;
